Update packages
And fix broken examples
