+++ .github/workflows/test.yml
@@ -1,13 +1,15 @@
-name: npm test
 on: push
+
 jobs:
-  setup-and-test:
+  test:
     runs-on: ubuntu-latest
     steps:
-      - uses: actions/checkout@v2
-      - uses: actions/setup-node@v2
+      - uses: actions/setup-node@v3
         with:
-          node-version: 12
-          check-latest: true
-      - run: npm install
-      - run: npm test
+          node-version: 18
+      - uses: pnpm/action-setup@v2
+        with:
+          version: latest
+      - uses: actions/checkout@v3
+      - run: pnpm i
+      - run: pnpm test
