+++ examples/localimports/package.json
@@ -1,13 +1,14 @@
 {
   "name": "localimports",
   "version": "0.0.0",
+  "type": "module",
   "scripts": {
     "dev": "vite",
     "build": "tsc && vite build",
     "serve": "vite preview"
   },
   "devDependencies": {
-    "typescript": "^4.2.4",
-    "vite": "^2.3.0"
+    "typescript": "^4.8.4",
+    "vite": "^3.2.3"
   }
 }
