+++ examples/multipage/index.html
@@ -9,6 +9,6 @@
   <body>
     <pug src="./index.pug"></pug>
     <div id="app"></div>
-    <script type="module" src="/src/main.ts"></script>
+    <script type="module" src="./src/main.ts"></script>
   </body>
 </html>
