+++ examples/multipage/package.json
@@ -1,13 +1,14 @@
 {
   "name": "multipage",
   "version": "0.0.0",
+  "type": "module",
   "scripts": {
     "dev": "vite",
     "build": "tsc && vite build",
     "serve": "vite preview"
   },
   "devDependencies": {
-    "typescript": "^4.2.3",
-    "vite": "^2.1.5"
+    "typescript": "^4.8.4",
+    "vite": "^3.2.3"
   }
 }
