+++ examples/multipage/src/main.ts
@@ -1,6 +1,6 @@
-import './style.css'
+import "./style.css"
 
-const app = document.querySelector<HTMLDivElement>('#app')!
+const app = document.querySelector<HTMLDivElement>("#app")!
 
 app.innerHTML = `
   <h1>Hello Vite!</h1>
