+++ examples/multipage/vite.config.ts
@@ -1,4 +1,3 @@
-import { resolve } from "path"
 import { defineConfig } from "vite"
 import pugPlugin from "../../src/index"
 
@@ -8,9 +7,9 @@ export default defineConfig({
   build: {
     rollupOptions: {
       input: {
-        main: resolve(__dirname, "index.html"),
-        nested: resolve(__dirname, "nested/index.html"),
-      },
-    },
-  },
+        main: "./index.html",
+        nested: "./nested/index.html"
+      }
+    }
+  }
 })
