--- jest.config.json
@@ -1,3 +1,0 @@
-{
-  "preset": "ts-jest"
-}
