+++ package.json
@@ -1,18 +1,38 @@
 {
   "name": "vite-plugin-pug",
-  "version": "0.3.1",
+  "version": "0.3.2",
   "description": "Vite plugin for transform Pug templates into HTML",
-  "main": "dist/index.js",
-  "module": "dist/index.mjs",
-  "types": "dist/index.d.ts",
+  "type": "module",
+  "main": "./dist/index.cjs",
+  "module": "./dist/index.js",
+  "types": "./dist/index.d.ts",
+  "exports": {
+    ".": {
+      "import": "./dist/index.js",
+      "require": "./dist/index.cjs"
+    }
+  },
   "files": [
     "dist"
   ],
+  "tsup": {
+    "entry": [
+      "src/index.ts"
+    ],
+    "format": [
+      "esm",
+      "cjs"
+    ],
+    "dts": true,
+    "clean": true,
+    "minify": true,
+    "sourcemap": true
+  },
   "scripts": {
     "dev": "npm run build -- --watch",
-    "build": "tsup src/index.ts --dts --format cjs,esm",
-    "typecheck": "tsc --noEmit",
-    "test": "jest"
+    "build": "tsup",
+    "lint": "tsc --noEmit",
+    "test": "vitest"
   },
   "repository": {
     "type": "git",
@@ -34,16 +54,10 @@
     "pug": "^3.0.2"
   },
   "devDependencies": {
-    "@types/jest": "^27.4.1",
     "@types/pug": "^2.0.6",
-    "jest": "^27.5.1",
-    "ts-jest": "^27.1.3",
-    "tsup": "^5.11.13",
-    "typescript": "^4.6.2",
-    "vite": "^2.8.3"
-  },
-  "peerDependencies": {
-    "vite": "~2",
-    "picocolors": "~1"
+    "tsup": "^6.4.0",
+    "typescript": "^4.8.4",
+    "vite": "^3.2.3",
+    "vitest": "^0.25.1"
   }
 }
