+++ tsconfig.json
@@ -1,18 +1,23 @@
 {
   "compilerOptions": {
-    "target": "ES2019",
-    "module": "CommonJS",
-    "lib": ["ES2019"],
-    "moduleResolution": "Node",
-    "declaration": true,
-    "declarationMap": false,
-    "sourceMap": false,
+    "baseUrl": ".",
+    "rootDir": "src",
     "outDir": "dist",
-    "strict": true,
-    "removeComments": true,
+    "target": "ESNext",
+    "module": "ESNext",
+    "moduleResolution": "Node",
+    "isolatedModules": true,
     "esModuleInterop": true,
+    "strict": true,
+    "allowJs": true,
+    "checkJs": true,
     "skipLibCheck": true,
+    "removeComments": true,
+    "noImplicitAny": false,
+    "noImplicitReturns": false,
+    "importsNotUsedAsValues": "error",
     "forceConsistentCasingInFileNames": true
   },
-  "include": ["src"]
+  "include": ["src/**/*"],
+  "exclude": ["node_modules", "dist", "examples"]
 }
