+++ vitest.config.ts
@@ -1,0 +1,9 @@
+import { defineConfig } from "vitest/config"
+
+export default defineConfig({
+  test: {
+    watch: false,
+    reporters: "verbose",
+    include: ["./test/**/*.{test,spec}.{js,ts}"]
+  }
+})
